Added control to toggle hoop
Added button that toggles hoop oscillation. WIll change the button depending on game later (hoop for plane, balloon speed for balloon game, etc). I also plan to change the button to a check box and add a slider. It also prints out the current oscilation value (true, false) and is saved in the vue object